requirements.txt
autopep8
black
coverage
cryptography
django-allauth
django-compressor
django-crispy-forms
django-nose
django-phonenumber-field
django-phonenumbers
flake8
pillow
pip-chill
psycopg2-binary
toml
types-cryptography
typing-extensions
